requirements.txt
# Core deep learning
torch>=2.0.1
torchvision>=0.15.2

# OpenCLIP (for ConvNeXt-XXL semantic features)
open-clip-torch>=2.24.0

# Image processing / augmentation
Pillow>=9.5.0
kornia>=0.7.2
kornia_rs>=0.1.2

# Data & metrics
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.2
scipy>=1.9.1

# Utilities
tqdm>=4.66.1
